Optimize SCM Transaction

+++ Modules/SupplyChain/Entities/StockMutation.php
@@ -84,11 +84,13 @@ class StockMutation extends MainModel
         parent::boot();
 
         static::deleting(function($StockMutation) {
-            foreach ($StockMutation->inbound_mutation_details as $inbound_mutation_detail) {
-                $inbound_mutation_detail->mutation_detail_initial_aging()->delete();
+            if (sizeOf($StockMutation->inbound_mutation_details) > 0) {
+                foreach ($StockMutation->inbound_mutation_details as $inbound_mutation_detail) {
+                    $inbound_mutation_detail->mutation_detail_initial_aging()->delete();
+                }
+                $StockMutation->inbound_mutation_details()->delete();
             }
-            $StockMutation->stock_mutation_details()->delete();
-            $StockMutation->item_stocks()->delete();
+            // $StockMutation->item_stocks()->delete();
             $StockMutation->approvals()->delete(); 
         });
     }
